Split DocumentationParserFunction
